// src/generated/pancake.ts
export interface SwapRequest {
  user:
    | UserContext
    | undefined;
  /** ERC-20 address */
  tokenIn: string;
  /** ERC-20 address */
  tokenOut: string;
  /** string to avoid JS bigint overflow */
  amountInWei: string;
  /** e.g. 50 = 0.5 % */
  slippageBps: number;
}

export interface SwapResponse {
  txHash: string;
  /** OK | FAILED */
  status: string;
  /** human-readable */
  error: string;
  /** machine code, e.g. INSUFFICIENT_GAS */
  code: string;
}

/** Allow privy server to sign for a user */
export interface UserContext {
  /** Ethereum wallet */
  address: string;
  /** Privy wallet-ID */
  walletId: string;
}

function createBaseSwapRequest(): SwapRequest {
  return { user: undefined, tokenIn: "", tokenOut: "", amountInWei: "", slippageBps: 0 };
}

export const SwapRequest: MessageFns<SwapRequest> = {
  encode(message: SwapRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.user !== undefined) {
      UserContext.encode(message.user, writer.uint32(10).fork()).join();
    }
    if (message.tokenIn !== "") {
      writer.uint32(18).string(message.tokenIn);
    }
    if (message.tokenOut !== "") {
      writer.uint32(26).string(message.tokenOut);
    }
    if (message.amountInWei !== "") {
      writer.uint32(34).string(message.amountInWei);
    }
    if (message.slippageBps !== 0) {
      writer.uint32(40).uint64(message.slippageBps);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SwapRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSwapRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.user = UserContext.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.tokenIn = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.tokenOut = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.amountInWei = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.slippageBps = longToNumber(reader.uint64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SwapRequest {
    return {
      user: isSet(object.user) ? UserContext.fromJSON(object.user) : undefined,
      tokenIn: isSet(object.tokenIn) ? globalThis.String(object.tokenIn) : "",
      tokenOut: isSet(object.tokenOut) ? globalThis.String(object.tokenOut) : "",
      amountInWei: isSet(object.amountInWei) ? globalThis.String(object.amountInWei) : "",
      slippageBps: isSet(object.slippageBps) ? globalThis.Number(object.slippageBps) : 0,
    };
  },

  toJSON(message: SwapRequest): unknown {
    const obj: any = {};
    if (message.user !== undefined) {
      obj.user = UserContext.toJSON(message.user);
    }
    if (message.tokenIn !== "") {
      obj.tokenIn = message.tokenIn;
    }
    if (message.tokenOut !== "") {
      obj.tokenOut = message.tokenOut;
    }
    if (message.amountInWei !== "") {
      obj.amountInWei = message.amountInWei;
    }
    if (message.slippageBps !== 0) {
      obj.slippageBps = Math.round(message.slippageBps);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SwapRequest>, I>>(base?: I): SwapRequest {
    return SwapRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SwapRequest>, I>>(object: I): SwapRequest {
    const message = createBaseSwapRequest();
    message.user = (object.user !== undefined && object.user !== null)
      ? UserContext.fromPartial(object.user)
      : undefined;
    message.tokenIn = object.tokenIn ?? "";
    message.tokenOut = object.tokenOut ?? "";
    message.amountInWei = object.amountInWei ?? "";
    message.slippageBps = object.slippageBps ?? 0;
    return message;
  },
};

function createBaseSwapResponse(): SwapResponse {
  return { txHash: "", status: "", error: "", code: "" };
}

export const SwapResponse: MessageFns<SwapResponse> = {
  encode(message: SwapResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.txHash !== "") {
      writer.uint32(10).string(message.txHash);
    }
    if (message.status !== "") {
      writer.uint32(18).string(message.status);
    }
    if (message.error !== "") {
      writer.uint32(26).string(message.error);
    }
    if (message.code !== "") {
      writer.uint32(34).string(message.code);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SwapResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSwapResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.txHash = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.status = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.error = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.code = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SwapResponse {
    return {
      txHash: isSet(object.txHash) ? globalThis.String(object.txHash) : "",
      status: isSet(object.status) ? globalThis.String(object.status) : "",
      error: isSet(object.error) ? globalThis.String(object.error) : "",
      code: isSet(object.code) ? globalThis.String(object.code) : "",
    };
  },

  toJSON(message: SwapResponse): unknown {
    const obj: any = {};
    if (message.txHash !== "") {
      obj.txHash = message.txHash;
    }
    if (message.status !== "") {
      obj.status = message.status;
    }
    if (message.error !== "") {
      obj.error = message.error;
    }
    if (message.code !== "") {
      obj.code = message.code;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SwapResponse>, I>>(base?: I): SwapResponse {
    return SwapResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SwapResponse>, I>>(object: I): SwapResponse {
    const message = createBaseSwapResponse();
    message.txHash = object.txHash ?? "";
    message.status = object.status ?? "";
    message.error = object.error ?? "";
    message.code = object.code ?? "";
    return message;
  },
};

function createBaseUserContext(): UserContext {
  return { address: "", walletId: "" };
}

export const UserContext: MessageFns<UserContext> = {
  encode(message: UserContext, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.address !== "") {
      writer.uint32(10).string(message.address);
    }
    if (message.walletId !== "") {
      writer.uint32(18).string(message.walletId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UserContext {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUserContext();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.address = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.walletId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UserContext {
    return {
      address: isSet(object.address) ? globalThis.String(object.address) : "",
      walletId: isSet(object.walletId) ? globalThis.String(object.walletId) : "",
    };
  },

  toJSON(message: UserContext): unknown {
    const obj: any = {};
    if (message.address !== "") {
      obj.address = message.address;
    }
    if (message.walletId !== "") {
      obj.walletId = message.walletId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UserContext>, I>>(base?: I): UserContext {
    return UserContext.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UserContext>, I>>(object: I): UserContext {
    const message = createBaseUserContext();
    message.address = object.address ?? "";
    message.walletId = object.walletId ?? "";
    return message;
  },
};
// ...
export interface SwapServiceServer extends UntypedServiceImplementation {
  executeSwap: handleUnaryCall<SwapRequest, SwapResponse>;
}
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}